Change setModalContent to use ajax
setModalContent in the achievments_overview.html file would previously
use data about Badges and BadgeRequests passed from the view.
Now it gets the relevant data from the new badge_request_data view,
which functions as an endpoint and provides data in a JsonResponse.

This has been done so that the request modal can be used in other
templates without altering the corresponding views, see following
commits.