Fix form rendering in request modal snippet
From the previous commit, the request modal snippet would attempt to
render a BadgeRequestForm using {{ form }}.
This only works if such a form is passed from the view with the proper
name, which was the case on the "Available Badges" page but not
otherwise.

This has been changed to include a manual rendering of the form in the
snippet file, so that the modal and form can be properly rendered
without passing a form.

This has been done so that the snippet can be used elsewhere without
altering the respective views.