Fix a broken datetime comparison
The comparison was between the self field "end" and "datetime.now"
There are two problems with this:

Firstly, the name "datetime" in that file is the top-level
datetime package, which has no method named "now".
Instead, the correct name of the method would be
"datetime.datetime.now". A likely reason for the mixup is that we
usually import the datetime module within the datetime package,
using "from datetime import datetime". When this is the case,
the method is found as "datetime.now".

The second issue was that the comparison was between "self.end",
which is a datetime object, and "datetime.now", which is a function.
When run, this would likely have returned a "NotImplemented".
The fix here is rather simple: just compare with "datetime.now()",
parentheses included. Now the comparison is between "self.end" and
the result of the "now"-function, which is also a datetime object.

Combining these two, the correct comparison becomes:
"datetime.datetime.now >= self.end"