Fix bug in update_scoreboards command
There was a bug in the update_scoreboards command that allowed hybrids
who are not longer active to remain on the "current" scoreboard for half
a year too long.

The problem was that checking if someone should be on the current
scoreboard was done with the following test:
`if hybrid.graduation_year >= current_year:`
During the fall semester of your graduation year you will already have
graduated, but the test will still allow you on the current scoreboard.

It has been replaced with the following line:
`if hybrid.get_grade() <= 5:`
This is better because get_grade takes the school year (rather than
calendar year) into account.