Redo kiltshop models file
This commit marks the beginning of a project to partially rewrite the
kiltshop app. In this commit I tackled the models.py-file, and made the
following changes:

Renamed model OrderInfo to OrderPeriod.
The name OrderInfo was misleading, as it seemed to describe info about a
single order. It has been changed to OrderPeriod which better reflects
what the model actually is: a period of time during which orders are
made. This brought with it some changes to other files: views, forms,
and admin. These files will be changed more later, and the
current changes to them are only to make sure this commit is
non-breaking.

Added a related_name to the OrderInfo-Order relation.
This relation had no related_name, and an ugly get_orderinfo function
was used instead. This was fixed, and the only usage (the
bestilling.html template) was amended to reflect the change.

General rewrites and PEP-8.
In addition to the above functional changes, several lines were
rewritten to be cleaner and more idiomatic. The file is now PEP-8
compliant.