Move OrderPeriod.orders relation to Order.period
Previously, the connection between an order and its period was through
the ManyToManyField OrderPeriod.orders. This was not ideal, as it meant
that an Order could belong to multiple periods.
This was changed to be a ForeignKey, Order.period. Since we set the
related_name to be orders, the relation is functionally the same from
OrderPeriod to Order. When going from Order to OrderPeriod, however, we
no longer have to deal with a ManyRelatedManager or call .first().

Most of the work to support this change was written in the migration.
It is a decent example of a combined schema- and data migration.