Fix error when creating new Order
After the last commit, an error was introduced. Since period was added
as a non-nullable field, attempting to create an Order without setting
this field would cause a crash. This happened when creating a new order
through the shop.

This should have been done as part of the previous commit, but I didn't
test enough scenarios before pushing.