Rename startTime and endTime on OrderPeriod
They were renamed to start_time and end_time to match PEP-8.
This should have been done in the models.py-commit, but I didn't catch
the error at that time.

Also, the variable names startTime and endTime were used a bit in
views.py. These have not been changed, as this is a change on the model
itself. Poor varible names in views will be adressed when views are
fixed in a later commit.