Rewrite order view
This view was not bad. Most of the changes in the view itself are small
cleanup changes, particularily to use relations in a more idimatic and
efficient manner.

The template was given a reformat for proper indenting, which changed
(almost?) every single line. Other changes were also implemented to
simplify the HTML itself and remove two unnecessary for-loops.
Functionality remains the same.

Finally, a change was made to the CSS-file to remove references to
broken tinypic-images that made selects wierd.

Things left to fix: HTML and CSS could probably do with a makeover, for
the visuals.