Add option to download order data as Excel file
In the new utils.py-file, a function has been added that takes the
primary key of an OrderPeriod object, and returns an Excel file
containing all the orders in that period. This requires the package
xlwt, which has been added to requirements.txt.

It is called from a new view under the URL
/kilt/admin/bestilling/<pk>/excel and the view download_period_as_excel.
Links to this has been added in both the list of periods and the page
that shows orders in a period.