Rewrite order_new and order_edit views
This one required quite a few changes.
Firstly, both views had form validation logic in them. This has been
moved to the form class itself. Also, it checked for empty form values
of start_time and end_time - these have now been made non-nullable in
the model itself.

These changes allowed the views to be wholly rewritten as class-based
views. The names were also changed to reflect the fact that they have to
do with OrderPeriods, not Orders.

The template was not changed majorly, mostly formatting.